# openapi/overseer-api.yaml
openapi: 3.0.3
info:
  title: Overseer API
  version: 5.4.1
  description: API canónica para observabilidade de pipelines e DAGs. Slack digest 08:30 Europe/Lisbon.
paths:
...
  /v1/catalog/pipelines/{pipeline_id}:
    patch:
      summary: Actualiza metadata (owner, schedule, etc.) com sync DB + YAML + SSH.
      parameters:
        - in: path
          name: pipeline_id
          required: true
          schema:
            type: string
  /v1/catalog/pipelines:
    post:
      summary: Regista ou atualiza o catálogo de um pipeline e o seu DAG.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pipeline_id
              properties:
                pipeline_id:
                  type: string
                name:
                  type: string
                owner:
                  type: string
                criticality:
                  type: string
                schedule:
                  type: string
                metadata:
                  type: object
                nodes:
                  type: array
                  items:
                    type: object
                    required:
                      - module_id
                    properties:
                      module_id:
                        type: string
                      label:
                        type: string
                      type:
                        type: string
                      metadata:
                        type: object
                edges:
                  type: array
                  items:
                    type: object
                    required:
                      - from_module_id
                      - to_module_id
                    properties:
                      from_module_id:
                        type: string
                      to_module_id:
                        type: string
                      metadata:
                        type: object
...